+++ assets/javascripts/kitten/components/menus/dashboard-menu/index.js
@@ -89,6 +89,7 @@ const StyledDashboardMenu = styled.nav`
     background-color: ${COLORS.font1};
     cursor: pointer;
   }
+  .k-DashboardMenu__expandable[open] .k-DashboardMenu__expandable__title,
   .k-DashboardMenu__expandable--hasActiveInside .k-DashboardMenu__expandable__title,
   .k-DashboardMenu__list > li > .k-DashboardMenu__item[aria-current="page"] {
     color: ${COLORS.background1};