+++ src/esm/index.js
@@ -140,7 +140,8 @@ export { List } from './components/lists/list';
 export { CheckList } from './components/lists/check-list';
 export { NavList } from './components/lists/nav-list';
 export { TagList } from './components/lists/tag-list';
-export { Timeline } from './components/lists/timeline'; // Loaders
+export { Timeline } from './components/lists/timeline';
+export { HorizontalTimeline } from './components/lists/horizontal-timeline'; // Loaders
 
 export { LoaderWithParagraph } from './components/loaders/loader-with-paragraph';
 export { Loader } from './components/loaders/loader'; // Menus
@@ -309,7 +310,10 @@ export { HomeIcon } from './components/icons/home-icon';
 export { TagIcon } from './components/icons/tag-icon';
 export { FileIcon } from './components/icons/file-icon';
 export { LoudspeakerIcon } from './components/icons/loudspeaker-icon';
-export { ShieldIcon } from './components/icons/shield-icon'; // ---------
+export { ShieldIcon } from './components/icons/shield-icon';
+export { PeopleIcon } from './components/icons/people-icon';
+export { StatsIcon } from './components/icons/stats-icon';
+export { SpeechBubbleIcon } from './components/icons/speech-bubble-icon'; // ---------
 // CONSTANTS
 // ---------