+++ src/helpers/utils/lazy-hook.js
@@ -19,8 +19,9 @@ var useLazyObserver = function useLazyObserver(lazyComponentRef) {
       withLazyObserver = _useState2[0],
       setLazyObserver = _useState2[1];
 
-  var hasNoRef = !(lazyComponentRef === null || lazyComponentRef === void 0 ? void 0 : lazyComponentRef.current);
   (0, _react.useEffect)(function () {
+    var hasNoRef = !(lazyComponentRef === null || lazyComponentRef === void 0 ? void 0 : lazyComponentRef.current);
+
     if (hasNoRef) {
       console.warn('lazyComponentRef.current does not exist, useLazyObserver will return true');
       setLazyObserver(true);