+++ src/index.js
@@ -773,6 +773,12 @@ Object.defineProperty(exports, "Timeline", {
     return _timeline.Timeline;
   }
 });
+Object.defineProperty(exports, "HorizontalTimeline", {
+  enumerable: true,
+  get: function get() {
+    return _horizontalTimeline.HorizontalTimeline;
+  }
+});
 Object.defineProperty(exports, "LoaderWithParagraph", {
   enumerable: true,
   get: function get() {
@@ -1667,6 +1673,24 @@ Object.defineProperty(exports, "ShieldIcon", {
     return _shieldIcon.ShieldIcon;
   }
 });
+Object.defineProperty(exports, "PeopleIcon", {
+  enumerable: true,
+  get: function get() {
+    return _peopleIcon.PeopleIcon;
+  }
+});
+Object.defineProperty(exports, "StatsIcon", {
+  enumerable: true,
+  get: function get() {
+    return _statsIcon.StatsIcon;
+  }
+});
+Object.defineProperty(exports, "SpeechBubbleIcon", {
+  enumerable: true,
+  get: function get() {
+    return _speechBubbleIcon.SpeechBubbleIcon;
+  }
+});
 Object.defineProperty(exports, "COLORS", {
   enumerable: true,
   get: function get() {
@@ -2046,6 +2070,8 @@ var _tagList = require("./components/lists/tag-list");
 
 var _timeline = require("./components/lists/timeline");
 
+var _horizontalTimeline = require("./components/lists/horizontal-timeline");
+
 var _loaderWithParagraph = require("./components/loaders/loader-with-paragraph");
 
 var _loader = require("./components/loaders/loader");
@@ -2337,6 +2363,12 @@ var _fileIcon = require("./components/icons/file-icon");
 var _loudspeakerIcon = require("./components/icons/loudspeaker-icon");
 
 var _shieldIcon = require("./components/icons/shield-icon");
+
+var _peopleIcon = require("./components/icons/people-icon");
+
+var _statsIcon = require("./components/icons/stats-icon");
+
+var _speechBubbleIcon = require("./components/icons/speech-bubble-icon");
 
 var _colorsConfig = _interopRequireDefault(require("./constants/colors-config"));